requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev.in
#
aiofiles==0.5.0           # via sanic
aiohttp==3.6.2            # via pytest-sanic
aioredis==1.3.1           # via insanic-framework
aiotask-context==0.6.1    # via insanic-framework
alabaster==0.7.12         # via sphinx
appdirs==1.4.4            # via virtualenv
async-generator==1.10     # via pytest-sanic
async-timeout==3.0.1      # via aiohttp, aioredis
attrs==20.2.0             # via aiohttp, pytest
babel==2.8.0              # via sphinx
bleach==3.2.1             # via readme-renderer
certifi==2020.6.20        # via httpx, requests
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via aiohttp, httpx, requests, zest.releaser
check-manifest==0.44      # via zest.releaser
click==7.1.2              # via pip-tools
colorama==0.4.3           # via twine, zest.releaser
contextvars==2.4          # via sniffio
coverage==5.3             # via -r tests.in, pytest-cov
distlib==0.3.1            # via virtualenv
docutils==0.16            # via pyroma, readme-renderer, sphinx
filelock==3.0.12          # via tox, virtualenv
h11==0.9.0                # via httpx
h2==3.2.0                 # via httpx
hiredis==1.1.0            # via aioredis
hpack==3.0.0              # via h2
hstspreload==2020.10.6    # via httpx
httptools==0.1.1          # via sanic
httpx==0.11.1             # via sanic
hyperframe==5.2.0         # via h2
identify==1.5.5           # via pre-commit
idna-ssl==1.1.0           # via aiohttp
idna==2.10                # via httpx, idna-ssl, requests, yarl
imagesize==1.2.0          # via sphinx
immutables==0.14          # via contextvars
importlib-metadata==2.0.0  # via importlib-resources, keyring, pep517, pluggy, pre-commit, pytest, tox, twine, virtualenv
importlib-resources==1.5.0  # via pre-commit, virtualenv
iniconfig==1.0.1          # via pytest
insanic-framework==0.9.1  # via -r tests.in
jinja2==2.11.2            # via sphinx
keyring==21.4.0           # via twine
markupsafe==1.1.1         # via jinja2
mirakuru==2.3.0           # via pytest-redis
multidict==4.7.6          # via aiohttp, sanic, yarl
nodeenv==1.5.0            # via pre-commit
packaging==20.4           # via bleach, pytest, sphinx, tox
pep517==0.8.2             # via check-manifest
pip-tools==5.3.1          # via -r dev.in
pkginfo==1.5.0.1          # via twine
pluggy==0.13.1            # via pytest, tox
port-for==0.4             # via pytest-redis
pre-commit==2.7.1         # via -r dev.in
prometheus-client==0.5.0  # via insanic-framework
psutil==5.7.2             # via insanic-framework, mirakuru
py==1.9.0                 # via pytest, tox
pygments==2.7.1           # via pyroma, readme-renderer, sphinx
pyjwt==1.7.1              # via insanic-framework
pyparsing==2.4.7          # via packaging
pyroma==2.6               # via zest.releaser
pytest-cov==2.10.1        # via -r tests.in
pytest-redis==2.0.0       # via -r tests.in
pytest-sanic==1.6.2       # via -r tests.in
pytest==6.1.1             # via -r tests.in, pytest-cov, pytest-redis, pytest-sanic
python-dateutil==2.8.1    # via insanic-framework
pytz==2020.1              # via babel
pyyaml==5.3.1             # via pre-commit
readme-renderer==26.0     # via twine, zest.releaser
redis==3.5.3              # via pytest-redis
requests-toolbelt==0.9.1  # via twine
requests==2.24.0          # via requests-toolbelt, sphinx, twine, zest.releaser
rfc3986==1.4.0            # via httpx, twine
sanic==20.6.3             # via -r tests.in, insanic-framework
six==1.15.0               # via bleach, packaging, pip-tools, python-dateutil, readme-renderer, tox, virtualenv, zest.releaser
sniffio==1.2.0            # via httpx
snowballstemmer==2.0.0    # via sphinx
sphinx-autodoc-typehints==1.11.0  # via -r docs.in
sphinx-issues==1.2.0      # via -r docs.in
sphinx-rtd-theme==0.5.0   # via -r docs.in
sphinx==3.2.1             # via -r docs.in, sphinx-autodoc-typehints, sphinx-issues, sphinx-rtd-theme, sphinxcontrib-log-cabinet
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-log-cabinet==1.0.1  # via -r docs.in
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
toml==0.10.1              # via check-manifest, pep517, pre-commit, pytest, tox
tox==3.20.1               # via -r tests.in
tqdm==4.50.0              # via twine
twine==3.2.0              # via zest.releaser
typing-extensions==3.7.4.3  # via aiohttp, yarl
ujson==4.0.1              # via sanic
urllib3==1.25.10          # via httpx, requests
uvloop==0.14.0            # via insanic-framework, sanic
virtualenv==20.0.33       # via pre-commit, tox
webencodings==0.5.1       # via bleach
websockets==8.1           # via sanic
wheel==0.35.1             # via zest.releaser
yarl==1.6.0               # via aiohttp
zest.releaser[recommended]==6.22.1  # via -r dev.in
zipp==3.3.0               # via importlib-metadata, importlib-resources, pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
